Rearrange some things
