+++ src/Compatibility/Core/src/AppHostBuilderExtensions.cs
@@ -176,6 +176,7 @@ namespace Microsoft.Maui.Controls.Hosting
 					Internals.Registrar.RegisterEffect("Xamarin", "ShadowEffect", typeof(ShadowEffect));
 #endif
 
+					// Update the mappings for ILabel/Label to work specifically for Controls
 					Label.RemapForControls();
 					
 				})
