+++ src/Controls/src/Core/HandlerImpl/Label/Label.Standard.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.Maui.Handlers;
+
+namespace Microsoft.Maui.Controls
+{
+	public partial class Label
+	{
+		public static void RemapForControls()
+		{
+		}
+
+		public static void MapTextType(LabelHandler handler, Label label) { }
+		public static void MapText(LabelHandler handler, Label label) { }
+	}
+}
