--- src/Controls/src/Core/Handlers/Label/ControlsLabelMappings.Standard.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Microsoft.Maui;
-using Microsoft.Maui.Controls;
-using Microsoft.Maui.Handlers;
-
-namespace Microsoft.Maui.Controls.Handlers
-{
-	public static partial class ControlsLabelMappings
-	{
-		public static void MapTextType(LabelHandler handler, Label label) { }
-		public static void MapText(LabelHandler handler, Label label) { }
-	}
-}
