--- src/Controls/src/Core/Handlers/Label/ControlsLabelMappings.Windows.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Microsoft.Maui.Handlers;
-
-namespace Microsoft.Maui.Controls.Handlers
-{
-	public static partial class ControlsLabelMappings
-	{
-		public static void MapTextType(LabelHandler handler, Label label) =>
-			Platform.TextBlockExtensions.UpdateText(handler.NativeView, label);
-
-		public static void MapText(LabelHandler handler, Label label) =>
-			Platform.TextBlockExtensions.UpdateText(handler.NativeView, label);
-	}
-}
